requirements.txt
adafruit-circuitpython-mcp230xx==2.4.5
paho-mqtt~=1.5.1
PyYAML==5.4.1
